rework opencv repositories
ibazel wants to run the query (buildfiles(deps(set(...)))), which
requires the local repository directories to exist. that doesn't work
well with our conditional opencv directories, as the mac one won't point
to a real directory when running on linux. instead we create a fake tree
that uses symlinks to find the actual directories.

this works on linux. hopefully mac too.